# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger",
        "version": "0.1"
    },
    "paths": {
...
        "/api/store": {
            "get": {
                "tags": [
                    "Store"
                ],
                "summary": "Get all stores of current user",
                "operationId": "69ab226e7ad5bc26e6f79dc000c98f81",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_online",
                        "in": "query",
                        "description": "Is the store online?",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StoreSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Server error"
                                        },
                                        "data": {
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Store"
                ],
                "summary": "Add store",
                "operationId": "1c859e1a0322fc4513a1f6ccec1003ca",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language (EN by default, or VN)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_online",
                        "in": "query",
                        "description": "Is the store online?",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/StoreSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Bad inputs"
                                        },
                                        "data": {
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Server error"
                                        },
                                        "data": {
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreSchema": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserSchema"
                    },
                    "name": {
                        "description": "Store name",
                        "type": "string",
                        "example": "Store name A",
                        "nullable": "false"
                    },
                    "Description": {
                        "description": "Description",
                        "type": "string",
                        "example": "Description A",
                        "nullable": "false"
                    },
                    "Address": {
                        "description": "Address",
                        "type": "string",
                        "example": "Address A",
                        "nullable": "false"
                    },
                    "is_online": {
                        "description": "Is online?",
                        "type": "boolean",
                        "example": true,
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
            "UserSchema": {
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "Kellen Boyer",
                        "nullable": "false"
                    },
                    "email": {
                        "description": "User E-mail",
                        "type": "string",
                        "example": "kellen.boyer@example.com",
                        "nullable": "false"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    }
}